@@ -16,7 +16,10 @@ import { changeRouterFactory } from './change/routes/changeRouter';
 export class ServerBuilder {
   private readonly serverInstance: express.Application;
 
-  public constructor(@inject(Services.CONFIG) private readonly config: IConfig, @inject(Services.LOGGER) private readonly logger: Logger) {
+  public constructor(
+    @inject(Services.CONFIG) private readonly config: IConfig,
+    @inject(Services.LOGGER) private readonly logger: Logger
+  ) {
     this.serverInstance = express();
   }
 
